refactor(transport): split monolithic transport.py into a package
src/tinyros/transport.py becomes src/tinyros/transport/ with one file
per concern, no behavior change:

- __init__.py     -- re-exports TinyServer, TinyClient (public).
                     Re-exports a small set of private symbols
                     (_recvall, _frame, _pack_oob, _unpack_oob,
                     _pack_call_large, _MSG_CALL_LARGE) that tests
                     reach into; not part of the public API.
- _common.py      -- module-level constants, _logger, env-var defaults
                     (_default_shm_threshold, _default_max_frame_bytes),
                     _PendingCall, _SENTINEL, the _INFLIGHT_ACQUIRE_POLL_S
                     constant, _LISTEN_BACKLOG.
- _framing.py     -- pure wire helpers: _recvall, _frame, _pack_oob /
                     _unpack_oob, the CALL_LARGE pack/parse/materialize
                     trio, _try_unlink_shm. No socket lifecycle or
                     threading state.
- _server.py      -- TinyServer (lifecycle, accept loop, reader loop,
                     in-flight semaphore, CALL_LARGE worker dispatch
                     including the drop-conn-on-corrupt-metadata path,
                     reply path).
- _client.py      -- TinyClient + _ClientMethod (lifecycle, send/recv
                     loops, encode_call, reconnect-on-send-failure
                     choreography under _pending_lock with the
                     drain-vs-call race fix and the oversized-reply
                     teardown protection, shm bookkeeping).

Public surface unchanged: 'from tinyros.transport import TinyServer,
TinyClient' works as before. Test imports of private symbols
('from tinyros.transport import _pack_oob' and friends) continue to
resolve via the package __init__.

Logger scope stays 'tinyros.transport' across all four submodules so
external observers see one logical component (per docs/standards/
code-organization.md's 'stop at a meaningful sub-component boundary'
rule).

docs/guides/architecture/overview.md is updated to reflect the new
layout and the per-file 'where to make changes' table.

Test files (tests/test_transport_framing.py, tests/test_transport_rpc.py)
are intentionally left where they are; mirroring the package
structure into tests/transport/* would be a separate refactor and is
not in scope here.
